Move authentication logic to concern
