+++ app/controllers/api/v1/application_controller.rb
@@ -4,25 +4,7 @@ module API
   module V1
     class ApplicationController < ActionController::API
       include API::V1::ErrorHandler
-
-      before_action :doorkeeper_authorize!
-
-      private
-
-      def current_user
-        User.find(doorkeeper_token.resource_owner_id) if doorkeeper_token
-      end
-
-      # :reek:FeatureEnvy
-      def doorkeeper_unauthorized_render_options(error: nil)
-        return unless error
-
-        {
-          json: {
-            errors: build_errors(detail: error.description, source: error.state, code: error.name)
-          }
-        }
-      end
+      include API::V1::DoorkeeperAuthentication
     end
   end
 end
